fix: update build-metrics import paths for routes/operational/ reorganization
The rebase onto main moved build-metrics.ts into routes/operational/
but left stale relative paths in the route file itself, its test, and
the crux client. This commit fixes those references.
